Updates tests to be run with nose.
